+++ cmd/amb-sidecar/filters/app/jwthandler/jwthandler.go
@@ -1,6 +1,7 @@
 package jwthandler
 
 import (
+	"context"
 	"net/http"
 	"strings"
 	"time"
@@ -11,9 +12,9 @@ import (
 	crd "github.com/datawire/apro/apis/getambassador.io/v1beta2"
 	"github.com/datawire/apro/cmd/amb-sidecar/filters/app/httpclient"
 	"github.com/datawire/apro/cmd/amb-sidecar/filters/app/middleware"
+	"github.com/datawire/apro/lib/filterapi"
 	"github.com/datawire/apro/lib/jwks"
 	"github.com/datawire/apro/lib/jwtsupport"
-	"github.com/datawire/apro/lib/util"
 )
 
 func inArray(needle string, haystack []string) bool {
@@ -25,29 +26,56 @@ func inArray(needle string, haystack []string) bool {
 	return false
 }
 
-type JWTHandler struct {
-	Filter crd.FilterJWT
+type JWTFilter struct {
+	Spec crd.FilterJWT
 }
 
-func (h *JWTHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logger := middleware.GetLogger(r.Context())
-	httpClient := httpclient.NewHTTPClient(logger, 0, h.Filter.InsecureTLS)
+func (h *JWTFilter) Filter(ctx context.Context, r *filterapi.FilterRequest) (filterapi.FilterResponse, error) {
+	logger := middleware.GetLogger(ctx)
+	httpClient := httpclient.NewHTTPClient(logger, 0, h.Spec.InsecureTLS)
 
-	token := strings.TrimPrefix(r.Header.Get("Authorization"), "Bearer ")
+	tokenString := strings.TrimPrefix(r.GetRequest().GetHttp().GetHeaders()["Authorization"], "Bearer ")
 
-	if err := validateToken(token, h.Filter, httpClient); err != nil {
-		logger.Infoln(err)
-		util.ToJSONResponse(w, http.StatusUnauthorized, &util.Error{Message: err.Error()})
-	} else {
-		w.WriteHeader(http.StatusOK)
+	token, err := validateToken(tokenString, h.Spec, httpClient)
+	if err != nil {
+		return middleware.NewErrorResponse(ctx, http.StatusUnauthorized, err, nil), nil
 	}
+
+	ret := &filterapi.HTTPRequestModification{}
+	for _, hf := range h.Spec.InjectRequestHeaders {
+		data := map[string]interface{}{
+			// "token" is intentionally similar to a
+			// *jwt.Token, but unwrapped a bit, since I
+			// don't want the jwt-go implementation to be
+			// part of our user-facing interface.
+			//
+			//"token": token,
+			"token": map[string]interface{}{
+				"Raw":       token.Raw,
+				"Header":    token.Header,
+				"Claims":    (map[string]interface{})(*(token.Claims.(*jwt.MapClaims))),
+				"Signature": token.Signature,
+			},
+		}
+		value := new(strings.Builder)
+		if err := hf.Template.Execute(value, data); err != nil {
+			return middleware.NewErrorResponse(ctx, http.StatusInternalServerError,
+				errors.Wrapf(err, "computing header field %q", hf.Name), nil), nil
+		}
+		ret.Header = append(ret.Header, &filterapi.HTTPHeaderReplaceValue{
+			Key:   hf.Name,
+			Value: value.String(),
+		})
+	}
+
+	return ret, nil
 }
 
-func validateToken(token string, filter crd.FilterJWT, httpClient *http.Client) error {
+func validateToken(signedString string, filter crd.FilterJWT, httpClient *http.Client) (*jwt.Token, error) {
 	jwtParser := jwt.Parser{ValidMethods: filter.ValidAlgorithms}
 
 	var claims jwt.MapClaims
-	_, err := jwtsupport.SanitizeParse(jwtParser.ParseWithClaims(token, &claims, func(t *jwt.Token) (interface{}, error) {
+	token, err := jwtsupport.SanitizeParse(jwtParser.ParseWithClaims(signedString, &claims, func(t *jwt.Token) (interface{}, error) {
 		if t.Method == jwt.SigningMethodNone && inArray("none", filter.ValidAlgorithms) {
 			return jwt.UnsafeAllowNoneSignatureType, nil
 		}
@@ -69,34 +97,34 @@ func validateToken(token string, filter crd.FilterJWT, httpClient *http.Client) 
 		return keys.GetKey(kid)
 	}))
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	now := time.Now().Unix()
 
 	if filter.RequireAudience || filter.Audience != "" {
 		if !claims.VerifyAudience(filter.Audience, filter.RequireAudience) {
-			return errors.Errorf("Token has wrong audience: token=%#v expected=%q", claims["aud"], filter.Audience)
+			return nil, errors.Errorf("Token has wrong audience: token=%#v expected=%q", claims["aud"], filter.Audience)
 		}
 	}
 
 	if filter.RequireIssuer || filter.Issuer != "" {
 		if !claims.VerifyIssuer(filter.Issuer, filter.RequireIssuer) {
-			return errors.Errorf("Token has wrong issuer: token=%#v expected=%q", claims["iss"], filter.Issuer)
+			return nil, errors.Errorf("Token has wrong issuer: token=%#v expected=%q", claims["iss"], filter.Issuer)
 		}
 	}
 
 	if !claims.VerifyExpiresAt(now, filter.RequireExpiresAt) {
-		return errors.New("Token is expired")
+		return nil, errors.New("Token is expired")
 	}
 
 	if !claims.VerifyIssuedAt(now, filter.RequireIssuedAt) {
-		return errors.New("Token used before issued")
+		return nil, errors.New("Token used before issued")
 	}
 
 	if !claims.VerifyNotBefore(now, filter.RequireNotBefore) {
-		return errors.New("Token is not valid yet")
+		return nil, errors.New("Token is not valid yet")
 	}
 
-	return nil
+	return token, nil
 }